updated new changes

+++ database/migrations/2022_05_18_203240_create_appointments_table.php
@@ -13,6 +13,10 @@ class CreateAppointmentsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('appointments')) {
+            return;
+        }
+
         Schema::create('appointments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade')->nullable();
